+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add AWSCNIPrefixDelegation annotation.
+
 ## [3.38.0] - 2021-11-12
 
 ### Added